converter, storage: extract block size conversion logic
Move Convert_v1_BlockSize_To_api_BlockIO and its helpers
(getOptimalBlockIO, getOptimalBlockIOForDevice, getDiscardGranularity,
getOptimalBlockIOForFile) from converter.go to storage/blockio.go
