+++ pkg/virt-launcher/virtwrap/converter/converter_test.go
@@ -383,41 +383,6 @@ var _ = Describe("Converter", func() {
 			Entry("on s390x", s390x, "virtio"),
 		)
 
-		DescribeTable("Should handle custom block sizes correctly per architecture", func(arch string, logical, physical uint, shouldSucceed bool) {
-			kubevirtDisk := &v1.Disk{
-				BlockSize: &v1.BlockSize{
-					Custom: &v1.CustomBlockSize{
-						Logical:            logical,
-						Physical:           physical,
-						DiscardGranularity: pointer.P(physical),
-					},
-				},
-			}
-			libvirtDisk := &api.Disk{}
-			err := Convert_v1_BlockSize_To_api_BlockIO(kubevirtDisk, libvirtDisk, arch)
-			if shouldSucceed {
-				Expect(err).ToNot(HaveOccurred())
-				expectedXML := fmt.Sprintf(`<Disk device="" type="">
-  <source></source>
-  <target></target>
-  <blockio logical_block_size="%d" physical_block_size="%d" discard_granularity="%d"></blockio>
-</Disk>`, logical, physical, physical)
-				data, xmlErr := xml.MarshalIndent(libvirtDisk, "", "  ")
-				Expect(xmlErr).ToNot(HaveOccurred())
-				Expect(string(data)).To(Equal(expectedXML))
-			} else {
-				Expect(err).To(MatchError(ContainSubstring("exceeds the maximum supported size")))
-			}
-		},
-			MultiArchEntry("valid 1234", uint(1234), uint(1234), true),
-			Entry("4096 on s390x", s390x, uint(4096), uint(4096), true),
-			Entry("2048 on s390x", s390x, uint(2048), uint(2048), true),
-			Entry("1024 on s390x", s390x, uint(1024), uint(1024), true),
-			Entry("8192 on s390x", s390x, uint(8192), uint(8192), false),
-			Entry("65536 on s390x", s390x, uint(65536), uint(65536), false),
-			Entry("1 MiB on s390x", s390x, uint(1048576), uint(1048576), false),
-		)
-
 		DescribeTable("should set sharable and the cache if requested", func(arch, expectedModel string) {
 			v1Disk := &v1.Disk{
 				Name: "mydisk",
@@ -1747,52 +1712,6 @@ var _ = Describe("Converter", func() {
 		},
 			MultiArchEntry(""),
 		)
-
-		Context("BlockIO", func() {
-			It("Should detect disk block sizes for a file DiskSource", func() {
-				v1Disk := v1.Disk{
-					Name: "test",
-					BlockSize: &v1.BlockSize{
-						MatchVolume: &v1.FeatureState{Enabled: pointer.P(true)},
-					},
-				}
-				apiDisk := api.Disk{Source: api.DiskSource{File: "/"}}
-				Expect(Convert_v1_BlockSize_To_api_BlockIO(&v1Disk, &apiDisk, amd64)).To(Succeed())
-
-				blockIO := apiDisk.BlockIO
-				Expect(blockIO.LogicalBlockSize).To(Equal(blockIO.PhysicalBlockSize))
-				// The default for most filesystems nowadays is 4096 but it can be changed.
-				// As such, relying on a specific value is flakey unless
-				// we create a disk image and filesystem just for this test.
-				// For now, as long as we have a value, the exact value doesn't matter.
-				Expect(blockIO.LogicalBlockSize).ToNot(BeZero())
-				Expect(blockIO.DiscardGranularity).ToNot(BeNil())
-				Expect(*blockIO.DiscardGranularity).To(Equal(blockIO.LogicalBlockSize))
-			})
-
-			It("Should fail for non-file or non-block devices", func() {
-				const blockIoConfigErrorMessage = "failed to configure disk with block size detection enabled"
-				v1Disk := v1.Disk{
-					Name: "test",
-					BlockSize: &v1.BlockSize{
-						MatchVolume: &v1.FeatureState{Enabled: pointer.P(true)},
-					},
-				}
-				apiDisk := api.Disk{Source: api.DiskSource{}}
-				Expect(Convert_v1_BlockSize_To_api_BlockIO(&v1Disk, &apiDisk, amd64)).To(MatchError(ContainSubstring(blockIoConfigErrorMessage)))
-			})
-
-			It("Should fail block size detection for a nil domain disk", func() {
-				const nilDiskErrorMessage = "disk is nil"
-				v1Disk := v1.Disk{
-					Name: "test",
-					BlockSize: &v1.BlockSize{
-						MatchVolume: &v1.FeatureState{Enabled: pointer.P(true)},
-					},
-				}
-				Expect(Convert_v1_BlockSize_To_api_BlockIO(&v1Disk, nil, amd64)).To(MatchError(ContainSubstring(nilDiskErrorMessage)))
-			})
-		})
 	})
 
 	Context("PCIe topology with NUMA alignment", func() {
