+++ pkg/virt-launcher/virtwrap/converter/storage/BUILD.bazel
@@ -3,6 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = [
+        "blockio.go",
         "devicenaming.go",
         "directio.go",
         "generated_mock_directio.go",
@@ -12,18 +13,23 @@ go_library(
     importpath = "kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/converter/storage",
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/pointer:go_default_library",
+        "//pkg/safepath:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/virt-launcher/virtwrap/api:go_default_library",
+        "//pkg/virt-launcher/virtwrap/disksource:go_default_library",
         "//pkg/virtiofs:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/log:go_default_library",
         "//vendor/go.uber.org/mock/gomock:go_default_library",
+        "//vendor/golang.org/x/sys/unix:go_default_library",
     ],
 )
 
 go_test(
     name = "go_default_test",
     srcs = [
+        "blockio_test.go",
         "devicenaming_test.go",
         "directio_test.go",
         "storage_suite_test.go",
@@ -33,6 +39,7 @@ go_test(
     deps = [
         ":go_default_library",
         "//pkg/libvmi:go_default_library",
+        "//pkg/pointer:go_default_library",
         "//pkg/virt-launcher/virtwrap/api:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/testutils:go_default_library",
